fix(admin-web): unblock Next build under workspace setup
Three issues surfaced by the first CI run, all on the admin-web
Docker build's `next build` step:

1. validateDirectory's `await import("node:fs/promises")` was
   reachable from `app/gas/page.tsx` via the package's barrel
   export. Webpack walks transitive imports for the client bundle
   and rejects `node:` schemes — failed with `UnhandledSchemeError:
   Reading from "node:fs/promises" is not handled by plugins`.

   Fix: split shared/gas-validation into two entry points.
   - `@silencer/gas-validation` — pure JS, browser-safe (schemas,
     validateAll, validateFile, references, error types).
   - `@silencer/gas-validation/node` — fs-backed validateDirectory
     only. Imports `node:fs/promises` and `node:path` at module
     top-level (was a runtime dynamic import; same effect for
     webpack which resolves the string statically).

   New file: `shared/gas-validation/validate-fs.ts`. Package's
   `exports` map gets a `./node` entry. The CLI's local handler
   now imports from `@silencer/gas-validation/node`. Admin's
   `app/gas/page.tsx` and `components/GasMonacoEditor.tsx` keep
   importing from the bare specifier — they only need
   `GAS_SCHEMAS`, which lives in the pure entry.

2. `outputFileTracingRoot` at the top level of next.config.mjs
   triggered `Unrecognized key(s) in object: 'outputFileTracingRoot'`.
   Next 14.2 still has it under `experimental.outputFileTracingRoot`;
   it was promoted to top-level in 15+. Moved under experimental.

3. `monaco-editor` is a peerDependency of `@monaco-editor/react`,
   resolved transitively in the previous per-package bun.lock but
   not auto-installed by the new workspace install + `--filter`
   pattern. The import in `GasMonacoEditor.tsx` is type-only, so
   tsc fails with `Cannot find module 'monaco-editor'`. Added it
   as an explicit dep on web/admin (^0.55.0 — matches the version
   the old lockfile resolved).

Verified: `bun run build` from web/admin completes successfully —
webpack compile + type-check + 17 routes generated.
