+++ web/admin/package.json
@@ -14,6 +14,7 @@
     "@types/dagre": "^0.7.54",
     "dagre": "^0.8.5",
     "fflate": "^0.8.2",
+    "monaco-editor": "^0.55.0",
     "next": "^14.2.4",
     "pako": "^2.1.0",
     "react": "^18.3.1",
